pyproject.toml
[dependency-groups]
dev = [
  "deptry"
]

[project]
authors = [
  {name = "Francesco Frassinelli", email = "francesco.frassinelli@nina.no"}
]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
  "django-environ",
  "dlt[duckdb]>=1.18.2",
  "enlighten>=1.14.1"
]
description = ""
license = "GPL-3.0+"
name = "traffikdata-sync"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[project.scripts]
traffikdata-sync = "main:start"

[tool.deptry.per_rule_ignores]
DEP002 = ["enlighten"]

[tool.ruff]
fix = true

[tool.ruff.lint]
ignore = ["COM812"]
select = ["E", "W", "I", "F", "UP", "S", "B", "A", "COM", "LOG", "PTH", "Q"]
